Support Securian's official font
